Add|Edit: getPomodoroCycle fun added in homeviewmodel, pomodoroControlSLider fixed float bug

+++ app/src/main/java/com/umutsaydam/zenfocus/presentation/home/PomodoroControlSlider.kt
@@ -63,7 +63,7 @@ fun PomodoroControlSlider(
                 )
             }
         )
-        Text(text = defaultSliderPosition.toString())
+        Text(text = defaultSliderPosition.toInt().toString())
 
         Button(
             modifier = Modifier
